added backend + first frontend code
