+++ README.md
@@ -1,9 +1,7 @@
 # findU
-
 Social app for finding nice people to discover things with
 
 # Frontend
-
 Creating a react project:
 cd root folder
 npm i
@@ -16,39 +14,37 @@ sudo apt-get install -y nodejs
 npm install
 vite --save-dev - Version 4.4.9 is in use
 
-# Frontend execution
-
+#  execution
 ls frontend
 npm install
 npm run dev
 
-# Backend
+#  Apollo Client Installation
+npm install @apollo/client graphql
 
+# Backend
 Installtion Apollo Server:
 npm install @apollo/server graphql
 npm run build -> building the GraphQL js files in dist folder
 npm run start -> starting the graphQL-Servers
+IP-in der Datenbank hinterlegen
 
 # Information
-
 PW:
 KfW: test
 Windows-local: test123
 Mac-loca: xita145
 
 # Setup
-
 Testing typescript locally:
 npm i -g typescript
 npm i --save-dev @types/node
 tsc --version
 
 # VSCode Setup
-
 1. Prettier: Format Changes to Prettier
 2. Oh my zsh:
    sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
 
 # Useful Keybinds
-
 1. Control + D & Tastenkombi: Markiert mehrere Zeilen und ändert Values
