+++ frontend/package.json
@@ -10,7 +10,8 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@apollo/server": "^4.9.3",
+    "@apollo/client": "^3.8.6",
+    "@apollo/server": "^4.9.4",
     "apollo-server-express": "^3.12.1",
     "bootstrap": "^5.3.2",
     "graphql": "^16.8.1",
